Copilot feedback

+++ tests/NATS.Client.ObjectStore.Tests/ObjectStoreTest.cs
@@ -77,12 +77,12 @@ public class ObjectStoreTest
             var data = await store.GetInfoAsync("k1", cancellationToken: cancellationToken);
 
             var sha = Base64UrlEncoder.Encode(SHA256.HashData(buffer));
-            var size = (ulong)buffer.Length;
+            var size = buffer.Length;
             var chunks = Math.Ceiling(size / 10.0);
 
             Assert.Equal($"SHA-256={sha}", data.Digest);
             Assert.Equal(chunks, data.Chunks);
-            Assert.Equal(size, data.Size);
+            Assert.Equal((ulong)size, data.Size);
         }
 
         // buffer with smaller last chunk
@@ -97,12 +97,12 @@ public class ObjectStoreTest
             var data = await store.GetInfoAsync("k2", cancellationToken: cancellationToken);
 
             var sha = Base64UrlEncoder.Encode(SHA256.HashData(buffer));
-            var size = (ulong)buffer.Length;
+            var size = buffer.Length;
             var chunks = Math.Ceiling(size / 10.0);
 
             Assert.Equal($"SHA-256={sha}", data.Digest);
             Assert.Equal(chunks, data.Chunks);
-            Assert.Equal(size, data.Size);
+            Assert.Equal((ulong)size, data.Size);
         }
 
         // Object name checks
@@ -195,7 +195,7 @@ public class ObjectStoreTest
         await store.PutAsync("k1", new byte[] { 65, 66, 67 }, cancellationToken);
 
         var info = await store.GetInfoAsync("k1", cancellationToken: cancellationToken);
-        Assert.Equal(3M, info.Size);
+        Assert.Equal(3UL, info.Size);
 
         var bytes = await store.GetBytesAsync("k1", cancellationToken);
         Assert.Equal(bytes, new byte[] { 65, 66, 67 });
@@ -207,8 +207,8 @@ public class ObjectStoreTest
 
         var info2 = await store.GetInfoAsync("k1", showDeleted: true, cancellationToken: cancellationToken);
         Assert.True(info2.Deleted);
-        Assert.Equal(0M, info2.Size);
-        Assert.Equal(0M, info2.Chunks);
+        Assert.Equal(0UL, info2.Size);
+        Assert.Equal(0, info2.Chunks);
         Assert.Equal(string.Empty, info2.Digest);
 
         // Put again
